+++ RELEASENOTES.md
@@ -3,6 +3,7 @@
 ### 4.3.0 (UNRELEASED)
 * Added feature for SPAD playback `spad_option`.
 * Added the ability to delete groups. [#680]
+* Added the ability to edit a single shift for a volunteer. [#824]
 * Fix for incorrect calculation of total and missed calls metrics on reports page.
 * Fix for scenario when `sms_combine` is enabled and `sms_ask` is disabled. [#706]
 * Fix to check that if postal code gathering is disabled and a service body override it set, that the service body override takes precedence. [#818]